# perf(hackathon): MarshalJSON proto generator