--- a/generated/storage/active_component_json.pb.go
+++ b/generated/storage/active_component_json.pb.go
@@ -22,6 +22,7 @@ func (m *ActiveComponent) MarshalJSON() ([]byte, error) {
 	str := buf.String()
 	return unsafe.Slice(unsafe.StringData(str), len(str)), nil
 }
+
 func (m *ActiveComponent) marshalJSON(buf *strings.Builder) (err error) {
 	if m == nil {
 		buf.WriteString("null")
@@ -126,6 +127,7 @@ func (m *ActiveComponent_ActiveContext) MarshalJSON() ([]byte, error) {
 	str := buf.String()
 	return unsafe.Slice(unsafe.StringData(str), len(str)), nil
 }
+
 func (m *ActiveComponent_ActiveContext) marshalJSON(buf *strings.Builder) (err error) {
 	if m == nil {
 		buf.WriteString("null")