--- a/generated/storage/administration_usage_json.pb.go
+++ b/generated/storage/administration_usage_json.pb.go
@@ -22,6 +22,7 @@ func (m *SecuredUnits) MarshalJSON() ([]byte, error) {
 	str := buf.String()
 	return unsafe.Slice(unsafe.StringData(str), len(str)), nil
 }
+
 func (m *SecuredUnits) marshalJSON(buf *strings.Builder) (err error) {
 	if m == nil {
 		buf.WriteString("null")