--- a/generated/storage/api_token_json.pb.go
+++ b/generated/storage/api_token_json.pb.go
@@ -22,6 +22,7 @@ func (m *TokenMetadata) MarshalJSON() ([]byte, error) {
 	str := buf.String()
 	return unsafe.Slice(unsafe.StringData(str), len(str)), nil
 }
+
 func (m *TokenMetadata) marshalJSON(buf *strings.Builder) (err error) {
 	if m == nil {
 		buf.WriteString("null")