--- a/generated/storage/auth_machine_to_machine_json.pb.go
+++ b/generated/storage/auth_machine_to_machine_json.pb.go
@@ -22,6 +22,7 @@ func (m *AuthMachineToMachineConfig) MarshalJSON() ([]byte, error) {
 	str := buf.String()
 	return unsafe.Slice(unsafe.StringData(str), len(str)), nil
 }
+
 func (m *AuthMachineToMachineConfig) marshalJSON(buf *strings.Builder) (err error) {
 	if m == nil {
 		buf.WriteString("null")
@@ -113,6 +114,7 @@ func (m *AuthMachineToMachineConfig_Mapping) MarshalJSON() ([]byte, error) {
 	str := buf.String()
 	return unsafe.Slice(unsafe.StringData(str), len(str)), nil
 }
+
 func (m *AuthMachineToMachineConfig_Mapping) marshalJSON(buf *strings.Builder) (err error) {
 	if m == nil {
 		buf.WriteString("null")