--- a/generated/storage/auth_provider_json.pb.go
+++ b/generated/storage/auth_provider_json.pb.go
@@ -22,6 +22,7 @@ func (m *AuthProvider) MarshalJSON() ([]byte, error) {
 	str := buf.String()
 	return unsafe.Slice(unsafe.StringData(str), len(str)), nil
 }
+
 func (m *AuthProvider) marshalJSON(buf *strings.Builder) (err error) {
 	if m == nil {
 		buf.WriteString("null")
@@ -265,6 +266,7 @@ func (m *AuthProvider_RequiredAttribute) MarshalJSON() ([]byte, error) {
 	str := buf.String()
 	return unsafe.Slice(unsafe.StringData(str), len(str)), nil
 }
+
 func (m *AuthProvider_RequiredAttribute) marshalJSON(buf *strings.Builder) (err error) {
 	if m == nil {
 		buf.WriteString("null")