--- a/generated/storage/blob_json.pb.go
+++ b/generated/storage/blob_json.pb.go
@@ -22,6 +22,7 @@ func (m *Blob) MarshalJSON() ([]byte, error) {
 	str := buf.String()
 	return unsafe.Slice(unsafe.StringData(str), len(str)), nil
 }
+
 func (m *Blob) marshalJSON(buf *strings.Builder) (err error) {
 	if m == nil {
 		buf.WriteString("null")