--- a/generated/storage/cluster_init_json.pb.go
+++ b/generated/storage/cluster_init_json.pb.go
@@ -22,6 +22,7 @@ func (m *InitBundleMeta) MarshalJSON() ([]byte, error) {
 	str := buf.String()
 	return unsafe.Slice(unsafe.StringData(str), len(str)), nil
 }
+
 func (m *InitBundleMeta) marshalJSON(buf *strings.Builder) (err error) {
 	if m == nil {
 		buf.WriteString("null")