--- a/generated/storage/common_json.pb.go
+++ b/generated/storage/common_json.pb.go
@@ -22,6 +22,7 @@ func (m *KeyValuePair) MarshalJSON() ([]byte, error) {
 	str := buf.String()
 	return unsafe.Slice(unsafe.StringData(str), len(str)), nil
 }
+
 func (m *KeyValuePair) marshalJSON(buf *strings.Builder) (err error) {
 	if m == nil {
 		buf.WriteString("null")