… generated/storage/cluster_json.pb.go
… generated/storage/compliance_config_json.pb.go
… generated/storage/compliance_integration_json.pb.go
… generated/storage/compliance_json.pb.go
… generated/storage/compliance_operator_json.pb.go
… generated/storage/compliance_operator_v2_json.pb.go
… generated/storage/config_json.pb.go
… generated/storage/container_runtime_json.pb.go
… generated/storage/cve_json.pb.go
… generated/storage/declarative_config_health_json.pb.go
… generated/storage/delegated_registry_config_json.pb.go
… generated/storage/deployment_json.pb.go
… generated/storage/discovered_cluster_json.pb.go
… generated/storage/external_backup_json.pb.go
… generated/storage/group_json.pb.go
… generated/storage/hash_json.pb.go
… generated/storage/helm_cluster_json.pb.go
… generated/storage/http_endpoint_json.pb.go
… generated/storage/image_component_json.pb.go
… generated/storage/image_integration_json.pb.go
… generated/storage/image_json.pb.go
… generated/storage/installation_json.pb.go
… generated/storage/integration_health_json.pb.go
… generated/storage/kube_event_json.pb.go
… generated/storage/labels_json.pb.go
… generated/storage/log_json.pb.go
… generated/storage/mitre_json.pb.go
… generated/storage/namespace_metadata_json.pb.go
… generated/storage/network_baseline_json.pb.go
… generated/storage/network_flow_json.pb.go