### Description

This PR creates JSON Marshaler generator for protobuf messages and store them in `_json.pb.go` files.
It's compatible (on test level – not on spec) with our `ProtoToJson` function but does not use reflection and allocates less memory.

```diff
- old.txt – ProtoToJSON
+ new.txt - MarshalJSON()
```

```
                                 │   old.txt    │               new.txt                │
                                 │    sec/op    │    sec/op     vs base                │
ProtoToJSON/storage.Alert-8        82.35µ ± 19%   24.73µ ± 15%  -69.97% (p=0.000 n=10)
ProtoToJSON/storage.Image-8        1.754m ± 11%   1.767m ± 14%        ~ (p=0.739 n=10)
ProtoToJSON/storage.Cluster-8      9.819µ ±  7%   5.004µ ± 10%  -49.05% (p=0.000 n=10)
ProtoToJSON/storage.Alert#01-8     71.66µ ±  8%   21.28µ ± 14%  -70.31% (p=0.000 n=10)
ProtoToJSON/storage.Deployment-8   28.57µ ± 14%   11.38µ ± 13%  -60.18% (p=0.000 n=10)
geomean                            78.08µ         35.05µ        -55.11%

                                 │    old.txt    │               new.txt                │
                                 │     B/op      │     B/op      vs base                │
ProtoToJSON/storage.Alert-8        42.955Ki ± 0%   8.167Ki ± 0%  -80.99% (p=0.000 n=10)
ProtoToJSON/storage.Image-8        1521.3Ki ± 0%   927.5Ki ± 0%  -39.03% (p=0.000 n=10)
ProtoToJSON/storage.Cluster-8       3.752Ki ± 0%   2.148Ki ± 0%  -42.74% (p=0.000 n=10)
ProtoToJSON/storage.Alert#01-8     35.696Ki ± 0%   6.839Ki ± 0%  -80.84% (p=0.000 n=10)
ProtoToJSON/storage.Deployment-8   10.593Ki ± 0%   3.716Ki ± 0%  -64.92% (p=0.000 n=10)
geomean                             39.21Ki        13.28Ki       -66.12%

                                 │   old.txt   │               new.txt               │
                                 │  allocs/op  │  allocs/op   vs base                │
ProtoToJSON/storage.Alert-8        422.00 ± 0%    91.00 ± 0%  -78.44% (p=0.000 n=10)
ProtoToJSON/storage.Image-8        2.898k ± 0%   2.254k ± 0%  -22.22% (p=0.000 n=10)
ProtoToJSON/storage.Cluster-8       38.00 ± 0%    15.00 ± 0%  -60.53% (p=0.000 n=10)
ProtoToJSON/storage.Alert#01-8     387.00 ± 0%    78.00 ± 0%  -79.84% (p=0.000 n=10)
ProtoToJSON/storage.Deployment-8   144.00 ± 0%    44.00 ± 0%  -69.44% (p=0.000 n=10)
geomean                             303.8         101.1       -66.73%
```

### User-facing documentation

- [ ] CHANGELOG is updated **OR** update is not needed
- [ ] [documentation PR](https://spaces.redhat.com/display/StackRox/Submitting+a+User+Documentation+Pull+Request) is created and is linked above **OR** is not needed

### Testing and quality

<!--
General Availability requirements: https://github.com/stackrox/stackrox/blob/master/PR_GA.md
Feature Flags usage: https://github.com/stackrox/stackrox/blob/master/pkg/features/README.md
-->

- [ ] the change is production ready: the change is GA or otherwise the functionality is gated by a feature flag
- [ ] CI results are inspected

#### Automated testing

<!--
If no tests have been contributed, please explain why unless it's obvious,
e.g., the PR is a one-line comment change.
-->

- [ ] added unit tests
- [ ] added e2e tests
- [ ] added regression tests
- [ ] added compatibility tests
- [ ] modified existing tests

#### How I validated my change

<!--
Use this space to explain **how you validated** that **your change functions
exactly how you expect it**. Feel free to attach JSON snippets, curl commands,
screenshots, etc. Apply a simple benchmark: would the information you provided
convince any reviewer or any external reader that you did enough to validate
your change.

It is acceptable to assume trust and keep this section light, e.g. as a
bullet-point list.

It is acceptable to skip testing in cases when CI is sufficient, or it's a
markdown or code comment change only. It is also acceptable to skip testing for
changes that are too taxing to test before merging. In such case you are
responsible for the change after it gets merged which includes reverting,
fixing, etc. Make sure you validate the change ASAP after it gets merged or
explain in PR when the validation will be performed. Explain here why you
skipped testing in case you did so.

Have you created automated tests for your change? Explain here which validation
activities you did manually and why so.
-->

change me!